+++ Vault-Coder/src/utopia/vault/coder/model/scala/declaration/ClassDeclaration.scala
@@ -3,7 +3,7 @@ package utopia.vault.coder.model.scala.declaration
 import utopia.vault.coder.model.scala.code.Code
 import utopia.vault.coder.model.scala.Visibility.Public
 import utopia.vault.coder.model.scala.datatype.{Extension, GenericType}
-import utopia.vault.coder.model.scala.{DeclarationDate, Parameters, Visibility}
+import utopia.vault.coder.model.scala.{Annotation, DeclarationDate, Parameters, Visibility}
 
 /**
   * Used for declaring scala classes
@@ -14,9 +14,9 @@ case class ClassDeclaration(name: String, genericTypes: Seq[GenericType] = Vecto
                             constructionParams: Parameters = Parameters.empty, extensions: Vector[Extension] = Vector(),
                             creationCode: Code = Code.empty, properties: Vector[PropertyDeclaration] = Vector(),
                             methods: Set[MethodDeclaration] = Set(), nested: Set[InstanceDeclaration] = Set(),
-                            visibility: Visibility = Public, description: String = "", author: String = "",
-                            headerComments: Vector[String] = Vector(), since: DeclarationDate = DeclarationDate.today,
-                            isCaseClass: Boolean = false)
+                            visibility: Visibility = Public, annotations: Seq[Annotation] = Vector(),
+                            description: String = "", author: String = "", headerComments: Vector[String] = Vector(),
+                            since: DeclarationDate = DeclarationDate.today, isCaseClass: Boolean = false)
 	extends InstanceDeclaration
 {
 	override val keyword = {
@@ -33,8 +33,9 @@ case class ClassDeclaration(name: String, genericTypes: Seq[GenericType] = Vecto
 	override protected def makeCopy(visibility: Visibility, genericTypes: Seq[GenericType],
 	                                extensions: Vector[Extension], creationCode: Code,
 	                                properties: Vector[PropertyDeclaration], methods: Set[MethodDeclaration],
-	                                nested: Set[InstanceDeclaration], description: String, author: String,
+	                                nested: Set[InstanceDeclaration], annotations: Seq[Annotation],
+	                                description: String, author: String,
 	                                headerComments: Vector[String], since: DeclarationDate) =
 		ClassDeclaration(name, genericTypes, constructionParams, extensions, creationCode, properties, methods,
-			nested, visibility, description, author, headerComments, since, isCaseClass)
+			nested, visibility, annotations, description, author, headerComments, since, isCaseClass)
 }