+++ Vault-Coder/src/utopia/vault/coder/model/scala/declaration/TraitDeclaration.scala
@@ -3,7 +3,7 @@ package utopia.vault.coder.model.scala.declaration
 import utopia.vault.coder.model.scala.Visibility.Public
 import utopia.vault.coder.model.scala.code.Code
 import utopia.vault.coder.model.scala.datatype.{Extension, GenericType}
-import utopia.vault.coder.model.scala.{DeclarationDate, Visibility}
+import utopia.vault.coder.model.scala.{Annotation, DeclarationDate, Visibility}
 
 /**
   * Used for declaring traits
@@ -14,7 +14,8 @@ case class TraitDeclaration(name: String, genericTypes: Seq[GenericType] = Vecto
                             extensions: Vector[Extension] = Vector(),
                             properties: Vector[PropertyDeclaration] = Vector(),
                             methods: Set[MethodDeclaration] = Set(), nested: Set[InstanceDeclaration] = Set(),
-                            visibility: Visibility = Public, description: String = "", author: String = "",
+                            visibility: Visibility = Public, annotations: Seq[Annotation] = Vector(),
+                            description: String = "", author: String = "",
                             headerComments: Vector[String] = Vector(), since: DeclarationDate = DeclarationDate.today,
                             isSealed: Boolean = false)
 	extends InstanceDeclaration
@@ -28,8 +29,9 @@ case class TraitDeclaration(name: String, genericTypes: Seq[GenericType] = Vecto
 	override protected def makeCopy(visibility: Visibility, genericTypes: Seq[GenericType],
 	                                extensions: Vector[Extension], creationCode: Code,
 	                                properties: Vector[PropertyDeclaration], methods: Set[MethodDeclaration],
-	                                nested: Set[InstanceDeclaration], description: String, author: String,
+	                                nested: Set[InstanceDeclaration], annotations: Seq[Annotation],
+	                                description: String, author: String,
 	                                headerComments: Vector[String], since: DeclarationDate) =
-		TraitDeclaration(name, genericTypes, extensions, properties, methods, nested, visibility, description, author,
-			headerComments, since, isSealed)
+		TraitDeclaration(name, genericTypes, extensions, properties, methods, nested, visibility, annotations,
+			description, author, headerComments, since, isSealed)
 }